refactor: SceneTranscriptController 분리
